# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Water Pump Monitoring System (SWPMS) API",
    "description": "\n### PT Ascon Multi Pratama - Industrial IoT & Water Pump Monitoring Dashboard\n\nSWPMS is a centralized, real-time, and historical water pump monitoring backend platform.\nThis API provides full operational telemetry queries, dynamic site/area/device/sensor topology management, alarm rule evaluation, audit logging, CSV reporting, and controlled commanded pump ON/OFF power operations.\n\n#### Authentication\nUse the `/api/v1/auth/login` endpoint to obtain a JWT Bearer token.\nProvide the token in the header:\n`Authorization: Bearer <your_jwt_token>`\n\n#### Postman Import Instructions\nYou can import this entire API into Postman directly by choosing **Import -> Link** and pasting:\n`http://localhost:3000/docs/json` or downloading the OpenAPI specification from `/docs/json`.\n        ",
    "version": "1.1.0",
    "contact": {
      "name": "PT Ascon Multi Pratama SCADA & Automation Team",
      "url": "https://ascon.co.id"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/areas/{id}/sensors": {
      "get": {
        "summary": "Get Sensors Bound to Area",
        "tags": [
          "Areas"
        ],
        "description": "Returns sensors (e.g. water level, room ambient temperature) assigned to this area.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/assets/{id}/sensors": {
      "get": {
        "summary": "Get Sensors Bound to Asset",
        "tags": [
          "Assets & Pumps"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/sensors/": {
      "get": {
        "summary": "List Sensors",
        "tags": [
          "Sensors & Bindings"
        ],
        "description": "Returns all sensor instances with active bindings, metric mappings, and source devices.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            },
            "in": "query",
            "name": "page_size",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sort_by",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "in": "query",
            "name": "sort_order",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "site_id",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sensor_type",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Create Sensor Instance",
        "tags": [
          "Sensors & Bindings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "site_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "device_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "code": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 255
                  },
                  "sensor_type": {
                    "type": "string",
                    "enum": [
                      "FLOW_METER",
                      "PRESSURE_SENSOR",
                      "TEMPERATURE_SENSOR",
                      "DISTANCE_SENSOR",
                      "LEVEL_SENSOR",
                      "CURRENT_SENSOR",
                      "VIBRATION_SENSOR",
                      "OTHER"
                    ]
                  },
                  "metric_code": {
                    "type": "string",
                    "minLength": 1
                  },
                  "unit": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "site_id",
                  "code",
                  "name",
                  "sensor_type",
                  "metric_code"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/sensors/topology": {
      "get": {
        "summary": "Get Dynamic Plant Topology Model",
        "tags": [
          "Sensors & Bindings"
        ],
        "description": "Returns configuration-driven site -> areas -> assets with sensors and area_sensors as specified in PRD Section 20.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "site_id",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/sensors/{id}": {
      "get": {
        "summary": "Get Sensor Detail",
        "tags": [
          "Sensors & Bindings"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "summary": "Update Sensor",
        "tags": [
          "Sensors & Bindings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "site_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "device_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "code": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 255
                  },
                  "sensor_type": {
                    "type": "string",
                    "enum": [
                      "FLOW_METER",
                      "PRESSURE_SENSOR",
                      "TEMPERATURE_SENSOR",
                      "DISTANCE_SENSOR",
                      "LEVEL_SENSOR",
                      "CURRENT_SENSOR",
                      "VIBRATION_SENSOR",
                      "OTHER"
                    ]
                  },
                  "metric_code": {
                    "type": "string",
                    "minLength": 1
                  },
                  "unit": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Delete Sensor",
        "tags": [
          "Sensors & Bindings"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/sensors/{id}/binding": {
      "get": {
        "summary": "Get Sensor Binding",
        "tags": [
          "Sensors & Bindings"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "summary": "Bind Sensor to Pump or Area",
        "tags": [
          "Sensors & Bindings"
        ],
        "description": "Binds a sensor to either an asset or an area. Replaces previous active binding.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "asset_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "area_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "role": {
                    "type": "string"
                  },
                  "effective_from": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Unbind Sensor",
        "tags": [
          "Sensors & Bindings"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
...
}